move Reporter implementations to sync SM namespace in preparation to create the same for the async SM
