+++ source/Appccelerate.StateMachine.Facts/Appccelerate.StateMachine.Facts.csproj
@@ -24,6 +24,10 @@
     <ProjectReference Include="..\Appccelerate.StateMachine\Appccelerate.StateMachine.csproj" />
   </ItemGroup>
 
+  <ItemGroup>
+    <Folder Include="AsyncMachine\Reports\" />
+  </ItemGroup>
+
   <Import Project="..\Appccelerate.StateMachine.Facts.props" />  
   
 </Project>
