+++ source/Appccelerate.StateMachine.Facts/Machine/Reports/YEdStateMachineReportGeneratorTest.cs
@@ -23,7 +23,7 @@ namespace Appccelerate.StateMachine.Facts.Machine.Reports
     using System.IO;
     using FluentAssertions;
     using StateMachine.Machine;
-    using StateMachine.Reports;
+    using StateMachine.Machine.Reports;
     using Xunit;
 
     public class YEdStateMachineReportGeneratorTest
