+++ source/Appccelerate.StateMachine/Appccelerate.StateMachine.csproj
@@ -93,6 +93,10 @@
   <ItemGroup>
     <None Include="..\..\LICENSE" Pack="true" PackagePath="" />
   </ItemGroup>
+  <ItemGroup>
+    <Folder Include="AsyncMachine\Reports\" />
+    <Folder Include="Machine\Reports\" />
+  </ItemGroup>
 
   <!-- xml documentation file generation -->
   <PropertyGroup Condition="'$(Configuration)|$(Platform)'=='Release|AnyCPU'">
