+++ source/Appccelerate.StateMachine/Machine/Reports/CsvTransitionsWriter.cs
@@ -1,5 +1,5 @@
 //-------------------------------------------------------------------------------
-// <copyright file="CsvStatesWriter.cs" company="Appccelerate">
+// <copyright file="CsvTransitionsWriter.cs" company="Appccelerate">
 //   Copyright (c) 2008-2019 Appccelerate
 //
 //   Licensed under the Apache License, Version 2.0 (the "License");
@@ -16,37 +16,38 @@
 // </copyright>
 //-------------------------------------------------------------------------------
 
-namespace Appccelerate.StateMachine.Reports
+namespace Appccelerate.StateMachine.Machine.Reports
 {
     using System;
     using System.Collections.Generic;
     using System.IO;
     using System.Linq;
 
-    using Machine.States;
+    using States;
+    using Transitions;
 
     /// <summary>
-    /// Writes the states of a state machine to a stream as csv.
+    /// Writes the transitions of a state machine to a stream as csv.
     /// </summary>
     /// <typeparam name="TState">The type of the state.</typeparam>
     /// <typeparam name="TEvent">The type of the event.</typeparam>
-    public class CsvStatesWriter<TState, TEvent>
+    public class CsvTransitionsWriter<TState, TEvent>
         where TState : IComparable
         where TEvent : IComparable
     {
         private readonly StreamWriter writer;
 
         /// <summary>
-        /// Initializes a new instance of the <see cref="CsvStatesWriter&lt;TState, TEvent&gt;"/> class.
+        /// Initializes a new instance of the <see cref="CsvTransitionsWriter&lt;TState, TEvent&gt;"/> class.
         /// </summary>
         /// <param name="writer">The writer.</param>
-        public CsvStatesWriter(StreamWriter writer)
+        public CsvTransitionsWriter(StreamWriter writer)
         {
             this.writer = writer;
         }
 
         /// <summary>
-        /// Writes the specified states.
+        /// Writes the transitions of the specified states.
         /// </summary>
         /// <param name="states">The states.</param>
         public void Write(IEnumerable<IStateDefinition<TState, TEvent>> states)
@@ -55,31 +56,43 @@ namespace Appccelerate.StateMachine.Reports
 
             Guard.AgainstNullArgument("states", states);
 
-            this.WriteStatesHeader();
+            this.WriteTransitionsHeader();
 
             foreach (var state in states)
             {
-                this.ReportState(state);
+                this.ReportTransitionsOfState(state);
             }
         }
 
-        private void WriteStatesHeader()
+        private void WriteTransitionsHeader()
         {
-            this.writer.WriteLine("Source;Entry;Exit;Children");
+            this.writer.WriteLine("Source;Event;Guard;Target;Actions");
         }
 
-        private void ReportState(IStateDefinition<TState, TEvent> state)
+        private void ReportTransitionsOfState(IStateDefinition<TState, TEvent> state)
         {
-            var entry = string.Join(", ", state.EntryActions.Select(action => action.Describe()));
-            var exit = string.Join(", ", state.ExitActions.Select(action => action.Describe()));
-            var children = string.Join(", ", state.SubStates.Select(s => s.Id.ToString()));
+            foreach (var transition in state.TransitionInfos)
+            {
+                this.ReportTransition(transition);
+            }
+        }
+
+        private void ReportTransition(TransitionInfo<TState, TEvent> transition)
+        {
+            string source = transition.Source.ToString();
+            string target = transition.Target != null ? transition.Target.Id.ToString() : "internal transition";
+            string eventId = transition.EventId.ToString();
+
+            string guard = transition.Guard != null ? transition.Guard.Describe() : string.Empty;
+            string actions = string.Join(", ", transition.Actions.Select(action => action.Describe()));
 
             this.writer.WriteLine(
-                "{0};{1};{2};{3}",
-                state.Id,
-                entry,
-                exit,
-                children);
+                "{0};{1};{2};{3};{4}",
+                source,
+                eventId,
+                guard,
+                target,
+                actions);
         }
     }
 }
